[cli] move update helpers to client_common
Also undo include reordering
