+++ src/client/cli/formatter/yaml_formatter.cpp
@@ -14,10 +14,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
+#include <multipass/cli/yaml_formatter.h>
 
 #include <multipass/cli/client_common.h>
 #include <multipass/cli/format_utils.h>
-#include <multipass/cli/yaml_formatter.h>
 #include <multipass/utils.h>
 
 #include <multipass/format.h>
